requirements.txt
vnstock3
snowflake-connector-python
pandas
notebook